lesson2/pyproject.toml
[project]
name = "homework1"
version = "0.1.0"
description = ""
authors = [
    {name = "Alexander Kostromin",email = "ghosts.gh0ust1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.2,<3.0)",
    "scikit-learn (>=1.5,<2.0)",
    "dvc (>=3.50,<4.0)",
    "great-expectations (>=0.18,<1.0)",
    "pyyaml (>=6.0,<7.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)",
    "ruff (>=0.13.2,<0.14.0)",
    "isort (>=6.0.1,<7.0.0)",
    "pre-commit (>=4.3.0,<5.0.0)"
]
